Microservices/device-management/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
pika
fastapi-mqtt
